Increase speed and robustness of ingestion process
Removing _query_hash was causing deadlocks when trying to ingest
multiple nights' data simultaneously in different qsub jobs.
Also suspected that the way this was implemented was not safe
for parallel execution. Solution was to move these operations into
the same transaction as the main operation.

The approach to removing _query_hash seemed to be very expensive,
as matching and operating on _everything_ in the database. Instead,
limited using a WHERE statement, changing these from dominant to
negligible use of DB time.

Finally, we were occasionally encountering a neo4j TransientError
(at least before the _query_hash removal was sped up), so I added
retries to operations failing in this way.